+++ eido/exceptions.py
@@ -2,7 +2,12 @@
 
 from abc import ABCMeta
 
-_all__ = ["PathAttrNotFoundError", "EidoSchemaInvalidError", "EidoFilterError"]
+_all__ = [
+    "EidoFilterError",
+    "EidoSchemaInvalidError",
+    "EidoValidationError",
+    "PathAttrNotFoundError",
+]
 
 
 class EidoException(Exception):
@@ -30,3 +35,11 @@ class EidoFilterError(EidoException):
 
     def __init__(self, key):
         super(EidoFilterError, self).__init__(key)
+
+
+class EidoValidationError(EidoException):
+    """Object was not validated successfully according to schema."""
+
+    def __init__(self, message, errors_by_type):
+        super().__init__(message)
+        self.errors_by_type = errors_by_type